Rename and refactor load_user_data
